lightbox and button functionality
added functionality to my filter buttons that gets the category value from the data-category attribute and depending on if it matches the value the category chosen then it shows correct gallery items and hides the others. also added in functionality and styling for my lightboxes in certain scenarios the lightbox is hidden and on specific clicks such as enhancing the image reveals the lightbox.